Feat: 이벤트 페이지 연결
- 이벤트 페이지: Fragment -> Activity로 변경
- 홈 - 이벤트 페이지 연결
