+++ Neoul/app/src/main/java/com/example/neoul/di/appModule.kt
@@ -2,6 +2,7 @@ package com.example.neoul.di
 
 import com.example.neoul.presentation.main.brand.BrandViewModel
 import com.example.neoul.presentation.main.category.CategoryViewModel
+import com.example.neoul.presentation.main.home.EventViewModel
 import com.example.neoul.presentation.main.home.HomeViewModel
 import com.example.neoul.presentation.main.my.MyPageViewModel
 import com.example.neoul.presentation.main.my.MyPageFragment
@@ -16,6 +17,7 @@ val appModule = module {
     single { Dispatchers.Main }
 
     viewModel { HomeViewModel() }
+    viewModel { EventViewModel() }
     viewModel { CategoryViewModel() }
     viewModel { BrandViewModel() }
     viewModel { StoryViewModel() }
