+++ Neoul/app/src/main/res/layout/activity_event.xml
@@ -6,9 +6,11 @@
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_event"
+        android:layout_marginHorizontal="20dp"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"/>
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
