+++ Neoul/app/src/main/res/layout/item_event.xml
@@ -1,10 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="18dp"
+
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <ImageView
+        android:src="@drawable/banner_main"
+        android:scaleType="fitXY"
         android:id="@+id/img_banner"
         android:layout_width="match_parent"
         android:layout_height="164dp"
@@ -35,6 +39,7 @@
         android:text="너울에서 세상을 함께 바꾸어 보아요"
         android:textSize="14sp"
         android:textColor="@color/coolGray60"
+        android:paddingBottom="10dp"
         />
